requirements.txt
Flask
pytz
mysql-connector-python
